+++ database/migrations/2015_11_16_205641_create_pingxx_transfer_table.php
@@ -18,7 +18,7 @@ class CreatePingxxTransferTable extends Migration
             $table->string('openid');
             $table->string('transfer_id');
             $table->string('payment_id');
-            $table->decimal('amount', 11, 0);
+            $table->unsignedInteger('amount');
             $table->string('currency')->defualt('cny');
             $table->string('recipient');
             $table->string('transaction_no');
