+++ database/migrations/2015_11_16_205914_create_order_billing_table.php
@@ -16,7 +16,7 @@ class CreateOrderBillingTable extends Migration
             $table->increments('id');
             $table->integer('order_id');
             $table->integer('user_id');
-            $table->decimal('amount', 11, 0);
+            $table->unsignedInteger('amount');
             $table->string('resource_type');
             $table->integer('resource_id');
             $table->string('status');
